golangci-lint update and math/rand/v2
- some new revive rules
- removed max and min, now that we have built-ins for them
- added gosec
  - File permissions
  - HTTP servers ReadHeaderTimeout value (and a better metric listener)
- migrated to math/rand/v2 and improved secret key generator
  - Moved to math/rand/v2 everywhere
  - Use crypto/rand for the initial secret key
  - Added some nolint marks on rand.* functions

+++ internal/bookmarks/crypto.go
@@ -20,7 +20,7 @@ func EncryptID(id uint64, expires time.Time) (string, error) {
 	buf := make([]byte, 16)
 	now := uint64(expires.Unix())
 	binary.LittleEndian.PutUint64(buf[0:], now)
-	binary.LittleEndian.PutUint64(buf[8:], uint64(id))
+	binary.LittleEndian.PutUint64(buf[8:], id)
 
 	// Encrypt the packet. There's not need for complex IV since
 	// it's the right size to encrypt the initial packet.
